# sys/atomic_utils: Functions for atomic access