--- a/tests/bench_sys_atomic_utils/Makefile
+++ b/tests/bench_sys_atomic_utils/Makefile
@@ -1,0 +1,7 @@
+include ../Makefile.tests_common
+
+USEMODULE += xtimer
+USEMODULE += atomic_utils
+USEMODULE += test_utils_interactive_sync
+
+include $(RIOTBASE)/Makefile.include