--- a/tests/bench_sys_atomic_utils/Makefile.ci
+++ b/tests/bench_sys_atomic_utils/Makefile.ci
@@ -1,0 +1,3 @@
+BOARD_INSUFFICIENT_MEMORY := \
+    stm32f030f4-demo \
+    #