--- a/tests/sys_atomic_utils/Makefile
+++ b/tests/sys_atomic_utils/Makefile
@@ -1,0 +1,9 @@
+include ../Makefile.tests_common
+
+USEMODULE += atomic_utils
+USEMODULE += fmt
+USEMODULE += random
+USEMODULE += shell
+USEMODULE += xtimer
+
+include $(RIOTBASE)/Makefile.include