--- a/tests/sys_atomic_utils/Makefile.ci
+++ b/tests/sys_atomic_utils/Makefile.ci
@@ -1,0 +1,10 @@
+BOARD_INSUFFICIENT_MEMORY := \
+    arduino-duemilanove \
+    arduino-leonardo \
+    arduino-nano \
+    arduino-uno \
+    atmega328p \
+    nucleo-f031k6 \
+    nucleo-l011k4 \
+    stm32f030f4-demo \
+    #