--- a/tests/unittests/tests-atomic_utils/Makefile
+++ b/tests/unittests/tests-atomic_utils/Makefile
@@ -1,0 +1,1 @@
+include $(RIOTBASE)/Makefile.base