--- a/tests/unittests/tests-atomic_utils/Makefile.include
+++ b/tests/unittests/tests-atomic_utils/Makefile.include
@@ -1,0 +1,2 @@
+USEMODULE += atomic_utils
+USEMODULE += random # <-- used for input data to operate on