--- a/tests/unittests/tests-atomic_utils/tests-atomic_utils.h
+++ b/tests/unittests/tests-atomic_utils/tests-atomic_utils.h
@@ -1,0 +1,38 @@
+/*
+ * Copyright (C) 2020 Otto-von-Guericke-Universität Magdeburg
+ *
+ * This file is subject to the terms and conditions of the GNU Lesser
+ * General Public License v2.1. See the file LICENSE in the top level
+ * directory for more details.
+ */
+
+/**
+ * @addtogroup  unittests
+ * @{
+ *
+ * @file
+ * @brief       Unittests for the atomic util module
+ *
+ * @author      Marian Buschsieweke <marian.buschsieweke@ovgu.de>
+ */
+
+#ifndef TESTS_ATOMIC_UTILS_H
+#define TESTS_ATOMIC_UTILS_H
+
+#include "embUnit.h"
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+/**
+ * @brief   The entry point of this test suite.
+ */
+void tests_atomic_util(void);
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif /* TESTS_ATOMIC_UTILS_H */
+/** @} */