… sys/include/atomic_utils.h
… tests/bench_sys_atomic_utils/main.c
… tests/sys_atomic_utils/README.md
… tests/sys_atomic_utils/main.c
… tests/unittests/tests-atomic_utils/tests-atomic_utils.c