### Contribution description

This PR provides a set of functions to perform atomic accesses. These functions closely resemble the C11 atomic library functions, but are implemented more efficiently:

- Lock-free atomic read-modify-write operations are usually slower than just disabling IRQs, even on big fat STM32F7 CPUs. Still, the compiler will assume lock-free implementations to always be the fastest option. This assumption does not hold on bare metal embedded targets.
- If no lock-free implementation is used, the compiler will emit a library call. Just disabling IRQs inline is more efficient (in terms of CPU cycles and ROM)
- In special cases (e.g. only one thread is writing to a variable), only the write part of the read-modify-write operation needs to be atomically. This can be implemented cheaper than full atomic operations and is provided here.
    - This is useful to fix https://github.com/RIOT-OS/RIOT/issues/9882
- Under certain conditions (e.g. when IRQs are disabled), mixing atomic accesses with non-atomic accesses is possible and faster. This library allows for this.
- Atomic operations for setting and clearing bits are provided
    - Unlike `atomic_fetch_or(&mask, 1 << bit)` or `atomic_fetch_and(&mask, ~(1 << bit))` this can be implemented lock-free and efficiently with bit-banding on Cortex M3 / M4 MCUs

### Testing procedure

Note: This currently only works for CortexM MCUs.

#### Performance Check
- Run the provided benchmark in `tests/bench_atomic_util`
- It contains a `README.md`. If the given documentation is not sufficient, I really should fix it there rather than here

##### Arithmetic Correctness Check
- Run the added unit tests: `BOARD=<SOME_CORTEX_M_BOARD> make -C tests/unittests/ tests-atomic_utils flash test`
    - This will be most interesting for Cortex M3, M4, and M4F boards

#### Testing "Atomic-ness"

- Run the test in `tests/sys_atomic_utils`
    - You can use `make test`, but you might want to manually test as well
    - The `README.md` and the in-application documentation should be sufficient. If not, I really should fix it there rather than here.

### Issues/PRs references

Useful for https://github.com/RIOT-OS/RIOT/issues/9882