@kaspar030 2020-06-22
Nice one!

Should we already add atomic single bit functions? `atomic_set_bit_u32(uint32_t *var, bit)` etc.
We can then easily hook them up to use bit banding if a platform supports it.