@maribu 2020-06-24
> We can then easily hook them up to use bit banding if a platform supports it.

Great idea. Do you have any good documentation for this? E.g. the few examples for using this I found that didn't use the bit-banding periph region did never allocate memory in the SRAM bit-banding region, but just created a wild pointer pointing somewhere in there. That might not be the best idea...

I think there should be some `__attribute__((region("bit-banding")))` magic or similar to make sure a variable actually gets placed in there. But I didn't find anything the like. If such exists, the API should also provide an attribute for bitmasks, which by default is an empty preprocessor macro.

The alternative would be to check within every `atomic_set_bit_u<WIDTH>()` function if the address of the target mask is within the SRAM bit banding region and else fall back to `atomic_fetch_or()`. But I much rather add a precondition to the function that bitmask are allocated with the magic keyword, so that users get predictable and consistent performance.