@maribu 2020-06-24
I think it is `__attribute__((section(''.srambb'')))` :-) Let me check if this works without touching the linker scripts