@maribu 2020-06-24
@kaspar030: I implemented the atomic set_bit / clear_bit functions as suggested using bit banding. Thanks for pointing me to this.

I added some unit tests for arithmetic correctness of the atomic functions; which directly has proven to be worth the effort.