@maribu 2020-06-27
I added one more test, which currently is only applicable to the `atomic_fetch_<OP>_u<BITS>()` family of functions. It also contains a broken implementation `volatile_fetch_<OP>_u<BITS>()` for reference, which can also be tested. The test seems to perform quite well to detect issues with the `volatile_fetch_<OP>_u<BITS>()` functions and does not detect any issues for the `atomic_*()` family of functions :-)

I still need to extend the test to also allow checking the `atomic_set_bit_u<BITS>()` and `atomic_clear_bit_u<BITS>()` functions. Also, some documentation will be needed for the test to be actually useful.