@maribu 2020-09-29
Now that it compiles on Murdock, let's also address Travis comments, and then finally squash.