@kaspar030 2020-09-30
I'm looking at the benchmark results, ran on nrf52840dk. Overall this seems like a nice speedup over the c11 atomics.

Some things I don't understand:

```
| atom |     load |    8 |          7815 µs |         15627 µs |         10939 µs |
| atom |     load |   16 |          7815 µs |         10940 µs |         15627 µs |
| atom |     load |   32 |          7814 µs |         15626 µs |         10939 µs |
| atom |     load |   64 |          9377 µs |         10939 µs |         28127 µs |
```

Somehow the numbers look weird. I'd expect 64b to take longer than 32bit, for example.

```
| atom |      add |    8 |         12502 µs |         10940 µs |         25002 µs |                                                                
| atom |      add |   16 |         18752 µs |         12502 µs |         21877 µs |                                                                
| atom |      add |   32 |         10939 µs |         12502 µs |         26565 µs |                                                                
| atom |      add |   64 |         17190 µs |         10939 µs |         57814 µs |                                                                
...
| semi |      add |    8 |         12502 µs |         14065 µs |         20315 µs |
| semi |      add |   16 |         12501 µs |         18752 µs |         25002 µs |
| semi |      add |   32 |         10939 µs |         14064 µs |         20314 µs |
| semi |      add |   64 |         17189 µs |         10939 µs |         57814 µs |
```

Isn't semi_atomic supposed to be less atomic or fall back to full atomic, and should thus be faster?