@maribu 2020-09-30
> I'd expect 64b to take longer than 32bit, for example.

Yes, it should. I will double check if I messed something up in the benchmark. Maybe it is also something like cache efficiency depending on the position in ROM issue - e.g. like the issue with the STM32F767ZI on which faster code ran slower, except when the ISR vectors where of different size.

> Isn't semi_atomic supposed to be less atomic or fall back to full atomic, and should thus be faster?

If lock-free writes are supported (up to word size usually) and the same corresponding fully atomic would contain locks, it will be implemented faster and lock free (but less atomic). Otherwise, it will just be an alias for the corresponding atomic operation (and should be just as fast). The semi-atomic operations taking longer indicates an issue with the benchmark, or other effects like cache-efficiency due to ROM position outweigh the actual performance differences.

I have the same hardware at home, I should be able to reproduce the number tonight. Hopefully, I just messed up the benchmark. It would be a pitty, if the numbers are unreliable.

(This is by the way a good use case for AVRs :-) If the numbers don't make sense there, I surely fucked up the benchmark. Being a stupidly simple and, hence, highly predictable platform has its upsides :-))