@maribu 2020-09-30
> (This is by the way a good use case for AVRs :-) If the numbers don't make sense there, I surely fucked up the benchmark. Being a stupidly simple and, hence, highly predictable platform has its upsides :-))

On AVR, only the lock-free functions are slightly faster (`atomic_store_u8()` and `atomic_load_u8()`), literally all other functions (`atomic_store_u16()` and up, `atomic_load_u16()` and up, and all other `atomic_fetch_<op>_u<width>()`) implemented by `atomic_utils.h` took pretty much the same time.

The lock-free functions (which will also be lock free with real C11 atomics) take (+- some error) the same time for both atomic-utils and C11 atomics. So at least one thing is as expected.

I'm not so fluent with AVR assembly, but the ARM assembly makes sense. It seems like the machine code for functions in the atomic utils is pretty much what one would expect. I can't make any sense of the numbers now :-/