@maribu 2020-10-12
I amended the API for the bit clearing and setting, so that there is a dedicated type for a reference to a bit. This is useful to avoid computing the bit-band alias address of a given bit multiple times, as now the bit reference can be computed ahead of time and stored.