@kaspar030 2020-11-03
The benchmark seems borked.
I tried to isolate some numbers, so just chopped out everything but the semi add functions.
Previous full results (which were in between all the others):
```
| semi |      add |    8 |         14064 µs |         15627 µs |         26565 µs |
| semi |      add |   16 |         14064 µs |         20314 µs |         25001 µs |
| semi |      add |   32 |         10939 µs |         14064 µs |         26564 µs |
| semi |      add |   64 |         20314 µs |         15627 µs |         57814 µs |
```

In isolation:
```
Note: LOWER IS BETTER!

+------+----------+------+------------------+------------------+------------------+
| mode |       op | bits |         volatile |      atomic util |       c11 atomic |
+------+----------+------+------------------+------------------+------------------+
| semi |      add |    8 |         14064 µs |         18752 µs |         25002 µs |
| semi |      add |   16 |         12502 µs |         15627 µs |         20314 µs |
| semi |      add |   32 |         10939 µs |         14064 µs |         20314 µs |
| semi |      add |   64 |         20314 µs |         17190 µs |         59377 µs |
+------+----------+------+------------------+------------------+------------------+
```

Both are on an nrf52840dk.