@kaspar030 2020-11-04
Somehow the benchmark is still not stable.
This is an isolated test:

```
main(): This is RIOT! (Version: 2020.10-devel-1529-g30a1e-atomic_utils)
Note: LOWER IS BETTER!

+------+----------+------+------------------+------------------+------------------+
| mode |       op | bits |         volatile |      atomic util |       c11 atomic |
+------+----------+------+------------------+------------------+------------------+
| semi |      add |    8 |         12501 µs |         14064 µs |         26564 µs |
| semi |      add |   16 |         14064 µs |         15626 µs |         25002 µs |
| semi |      add |   32 |         10939 µs |         14064 µs |         20314 µs |
| semi |      add |   64 |         17189 µs |         23439 µs |         59377 µs |
+------+----------+------+------------------+------------------+------------------+
```

This is the same but together with the others:

```
| semi |      add |    8 |         18752 µs |         15627 µs |         20314 µs |
| semi |      add |   16 |         14064 µs |         15626 µs |         20314 µs |
| semi |      add |   32 |         10939 µs |         14065 µs |         20315 µs |
| semi |      add |   64 |         17189 µs |         23440 µs |         57815 µs |
```

We can write those differences off as different optimization opportunities by the compiler...